requirements.txt
ultralytics
fast-plate-ocr[onnx]
opencv-python
numpy
python-Levenshtein